+++ src/utils/index.ts
@@ -3,5 +3,6 @@ export * from "./date-utils";
 export * from "./file-operations";
 export * from "./frontmatter-utils";
 export * from "./note-utils";
+export * from "./period-navigation";
 export * from "./time-budget-utils";
 export * from "./validation-utils";